renamed/moved camera related screens into dedicated directory
