+++ lib/services/daily_photo_capture_service.dart
@@ -7,8 +7,8 @@ import 'package:life_frame/services/image_filesystem.dart';
 import 'package:life_frame/services/location.dart';
 import '../controllers/daily_entry_controller.dart';
 import '../models/daily_entry.dart';
-import '../screens/simple_camera_screen.dart';
-import '../screens/photo_confirmation_screen.dart';
+import '../screens/capture/camera_screen.dart';
+import '../screens/capture/photo_confirmation_screen.dart';
 import '../services/photo_stitching_service.dart';
 import 'package:talker/talker.dart';
 
@@ -147,7 +147,7 @@ class DailyPhotoCaptureService {
     final result = await Navigator.of(context).push<FramePhotos>(
       CupertinoPageRoute(
         fullscreenDialog: true,
-        builder: (context) => const SimpleCameraScreen(),
+        builder: (context) => const CameraScreen(),
       ),
     );
